main/TextBoxComponent optional/required immediate validate
- Pulled common validation methods into TextBoxComponentLike.
- main/TextBoxComponent.setValue now validates if value changed.
